+++ Interface/AddOns/AzeritePowerWeights/CHANGELOG.md
@@ -1,3 +1,11 @@
+### 8.3.11
+
+- Add Holy Priest Default-scale by Simbiawow
+
+### 8.3.10
+
+- Fix error in one of the lines in French translations causing error on login. This fixes [Issue #67](https://www.curseforge.com/wow/addons/azeritepowerweights/issues/67).
+
 ### 8.3.9
 
 - Update French translations (thanks to Simbiawow)